+++ tee-worker/identity/client-api/parachain-api/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Type definitions for the new OmniAccount Pallet.
 -   Expose getter to get the derived `omniaccount` address from a given identity
 -   Expose the OmniAccount `request_intent`'s trusted call.
+-   Add `TrustedCallAuthenticated` and `TCAuthentication` trusted call structs.
 
 ## [0.9.20-4.1] - 2024-09-30